backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
mysql-connector-python==8.1.0
SQLAlchemy==2.0.23
Werkzeug==2.3.7
python-dotenv==1.0.0
passlib==1.7.4
marshmallow==3.20.1
pymysql==1.1.0
cryptography==41.0.4
